requirements.txt
absl-py==1.3.0
astunparse==1.6.3
cachetools==4.2.4
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
einops==0.7.0
et-xmlfile==1.1.0
ffmpeg==1.4
fonttools==4.38.0
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.0
h5py==2.10.0
idna==3.4
importlib-metadata==5.0.0
joblib==1.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mkl-fft==1.3.0
numpy==1.20.0
oauthlib==3.2.2
onnx==1.7.0
onnxruntime==1.5.2
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==23.2
pandas==1.2.0
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.4.1
seaborn==0.13.0
sklearn==0.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==2.1.1
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.12
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
xlrd==2.0.1
zipp==3.10.0
